[IMP] ✨ More permissive ways of modifying and reopening contracts
