+++ energy_communities_service_invoicing/__manifest__.py
@@ -30,6 +30,7 @@
         "data/contract_line_qty_formula_data.xml",
         "data/product_data.xml",
         "views/contract_line_formula_views.xml",
+        "views/contract_template_views.xml",
         "views/contract_views.xml",
         "views/res_company_views.xml",
         "views/res_partner_views.xml",
