+++ energy_communities_service_invoicing/models/__init__.py
@@ -1,3 +1,4 @@
+from . import service_invoicing_info_mixin
 from . import abstract_contract
 from . import account_move
 from . import contract
