+++ energy_communities_service_invoicing/models/contract_line.py
@@ -13,6 +13,15 @@ class ContractLine(models.Model):
         for record in self:
             record.is_cancel_allowed = True
 
-    # TODO:
-    # skip last_date_invoice validation for modification action if contract is ready to start or active on free plan.
-    # def _check_last_date_invoiced(self):
+    # skip last_date_invoice update for modification action if contract is ready to start or on free plan.
+    def _update_recurring_next_date(self):
+        # TODO: Pay attention to original code in order to detect if method has been renamed:
+        # FIXME: Change method name according to real updated field
+        # e.g.: _update_last_date_invoiced()
+        for record in self:
+            if (
+                record.contract_id.status != "ready_to_start"
+                or record.contract_id.is_free_pack
+            ):
+                return
+        super()._update_recurring_next_date()
