+++ energy_communities_service_invoicing/models/res_company.py
@@ -4,7 +4,7 @@ from odoo.tools.translate import _
 
 class ResCompany(models.Model):
     _name = "res.company"
-    _inherit = "res.company"
+    _inherit = ["res.company"]
 
     service_invoicing_journal_id = fields.Many2one(
         comodel_name="account.journal",
