+++ energy_communities_service_invoicing/views/contract_template_views.xml
@@ -1,0 +1,12 @@
+<odoo>
+<record id="contract_template_form_view" model="ir.ui.view">
+  <field name="name">contract.template form view (in energy_communities service_invoicing)</field>
+  <field name="model">contract.template</field>
+  <field name="inherit_id" ref="contract.contract_template_form_view" />
+  <field name="arch" type="xml">
+    <xpath expr="//field[@name='contract_type']" position="after">
+      <field name="is_free_pack" />
+    </xpath>
+  </field>
+</record>
+</odoo>
